refactor: move error analyzer to own module
- seperate into smaller files
